fix: 프로필 이미지 수정 오류 해결

See merge request cloud_track/class_04/web_project3/team01/backend!88

+++ src/main/java/com/team01/billage/user/service/ProfileService.java
@@ -52,7 +52,7 @@ public class ProfileService {
 
         if (imageFile != null && !imageFile.isEmpty()) {
             // 기존 이미지가 있다면 삭제
-            if (imageUrl != null && !imageUrl.isEmpty()) {
+            if (imageUrl != null && !imageUrl.isEmpty() && !imageUrl.startsWith("/images")) {
                 s3BucketService.delete(imageUrl);
             }
             // 새 이미지 업로드
